requirements.txt
Package           Version  
----------------- ---------
asyncio           3.4.3    
backcall          0.2.0    
certifi           2021.10.8
chardet           4.0.0    
colorama          0.4.4    
debugpy           1.6.0    
decorator         5.1.1    
entrypoints       0.4      
fyers-apiv2       2.0.5    
idna              2.10     
ipykernel         6.13.0   
ipython           7.32.0   
jedi              0.18.1   
jupyter-client    7.2.2    
jupyter-core      4.10.0   
matplotlib-inline 0.1.3    
nest-asyncio      1.5.5    
numpy             1.21.6   
packaging         21.3     
pandas            1.3.5    
parso             0.8.3    
pickleshare       0.7.5    
pip               19.0.3   
prompt-toolkit    3.0.29   
psutil            5.9.0    
Pygments          2.11.2   
pyparsing         3.0.8    
python-dateutil   2.8.2    
pytz              2022.1   
pywin32           303      
pyzmq             22.3.0   
requests          2.25.1   
setuptools        40.8.0   
six               1.16.0   
tornado           6.1      
tqdm              4.64.0   
traitlets         5.1.1    
urllib3           1.26.9   
wcwidth           0.2.5    
websocket-client  1.2.1    
websockets        8.1      
